assembler: move new lexer, parser into separate files
